add github workflow
this makes a few of the fns on the `Queue` generic over a new `Playable`
trait to make the tests work without having to create an output stream.
while this is necessary as the github ci cannot open an output device,
which is why i am doing this change, it is also just a good idea in
general.