mpris/volume: round up for floating point impresicion when casting to u8
apparently due to floating point imprecision playerctl sometimes sends
me values like `0.649999999...` instead of `0.65`. this is then
multiplied by 100 and casted to `u8`, which always performs a truncating
cast, which results in `64` instead of `65`, which would make more sense.
so perform a `.ceil` first if the fract part is > 0.99, to avoid
truncating on floating point imprecision.